src/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
dnspython==2.1.0
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.1
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
whois==0.9.13
wrapt==1.12.1
